rename module to nccl
